# template-api/openapi.yaml
openapi: 3.0.3
info:
  title: FormGPT API
  version: "0.1.0"
  description: API для конструктора опросов с GPT помощником

servers:
  - url: http://survey-service:8080

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        createdAt:
          type: string
          format: date-time

    CreateSurveyRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
        description:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/CreateQuestionRequest"

    SurveyResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        ownerId:
          type: string
          format: uuid
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuestionResponse"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateQuestionRequest:
      type: object
      required:
        - title
        - type
      properties:
        id:
          type: integer
          nullable: true
          description: |
            ID существующего вопроса для обновления.
            Если null или отсутствует - создается новый вопрос.
        title:
          type: string
        type:
          type: string
          enum: [single-choice, multiple-choice, text, scale-1-10]
        options:
          type: array
          items:
            type: string

    QuestionResponse:
      type: object
      properties:
        id:
          type: integer
        surveyId:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [single-choice, multiple-choice, text, scale-1-10]
        options:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Answer:
      type: object
      properties:
        questionId:
          type: integer
        value:
          type: string

    SurveyWithAnswersResponse:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        ownerId:
          type: string
          format: uuid
        hasUserResponded:
          type: boolean
          description: Ответил ли текущий пользователь на этот опрос
        userRespondedAt:
          type: string
          format: date-time
          nullable: true
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuestionWithAnswerResponse"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    QuestionWithAnswerResponse:
      type: object
      properties:
        id:
          type: integer
        surveyId:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [single-choice, multiple-choice, text, scale-1-10]
        options:
          type: array
          items:
            type: string
        userAnswer:
          type: string
          description: Ответ текущего пользователя на этот вопрос
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    SurveyAnswer:
      type: object
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/Answer"

    GPTAnalysis:
      type: object
      properties:
        text:
          type: string
        score:
          type: string
          enum: [good, average, bad]

security:
  - bearerAuth: []

paths:
  /health:
    get:
      summary: Проверка работы сервиса
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /auth/register:
    post:
      summary: Регистрация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Успешная регистрация
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string

  /auth/login:
    post:
      summary: Логин
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Успешная авторизация
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string

  /users/me:
    get:
      summary: Профиль пользователя
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Данные пользователя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /surveys:
    get:
      summary: Список всех своих опросов
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Список опросов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SurveyResponse"

    post:
      summary: Создать новый опрос
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSurveyRequest"
      responses:
        '201':
          description: Опрос создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurveyResponse"

  /surveys/{surveyId}:
    get:
      summary: Получить конкретный опрос
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: surveyId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Опрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurveyWithAnswersResponse"

    put:
      summary: Обновить опрос
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: surveyId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSurveyRequest"
      responses:
        '200':
          description: Обновлённый опрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurveyResponse"

    delete:
      summary: Удалить опрос
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: surveyId
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Успешно удалён

  /surveys/{surveyId}/answers:
    post:
      summary: Отправка ответов на опрос
      description: |
        Отправляет ответы на опрос. 
        Если пользователь уже отвечал на этот опрос - его предыдущие ответы обновляются.
        
        🔍 **Проверки:**
        - Обязательные вопросы должны быть заполнены
        - Для single-choice ответ должен быть одним из предложенных вариантов
        - Для scale-1-10 ответ должен быть числом от 1 до 10
        - Для text ответ не должен превышать 1000 символов
        
        ✅ **Особенности:**
        - При повторной отправке ответов старые ответы заменяются новыми
        - вопросы можно пропускать (value = null)
        - Для multiple-choice передавайте варианты через `;`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: surveyId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SurveyAnswer"
      responses:
        '200':
          description: Ответы успешно сохранены или обновлены
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "success"

  /gpt/surveys/create:
    post:
      summary: Создание черновика опроса через GPT
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                targetAudience:
                  type: string
                questionCount:
                  type: integer
      responses:
        '200':
          description: Сгенерированный опрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurveyResponse"

  /gpt/surveys/{surveyId}/edit:
    put:
      summary: Улучшение опроса через GPT
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: surveyId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                prompt:
                  type: string
                  description: Необязательный пользовательский промт для улучшения
                  nullable: true
                  example: "Сделай вопросы более понятными для подростков"
      responses:
        '200':
          description: Улучшенный опрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurveyResponse"

  /gpt/questions/{questionId}/edit:
    put:
      summary: Улучшение вопроса через GPT
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: questionId
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                prompt:
                  type: string
                  description: Необязательный пользовательский промт для улучшения
                  nullable: true
                  example: "Сделай варианты ответа более разнообразными"
      responses:
        '200':
          description: Улучшенный вопрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionResponse"

  /gpt/surveys/{surveyId}/analyze:
    post:
      summary: Анализ качества опроса через GPT
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: surveyId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Рекомендации и оценка опроса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GPTAnalysis"

  /surveys/{surveyId}/stats:
    get:
      summary: Получение статистики по опросу
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: surveyId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Статистика опроса
          content:
            application/json:
              schema:
                type: object
                properties:
                  respondents:
                    type: integer
                  answersDistribution:
                    type: object